feat: simplify http version parsers
