+++ src/transport/http/binary_receiver.ts
@@ -2,8 +2,8 @@ import { CloudEvent, Version } from "../..";
 import { CloudEventV1, validateV1 } from "../../event/v1";
 import { CloudEventV03, validateV03 } from "../../event/v03";
 import { Headers, validate } from "./headers";
-import { binaryParsers as v1Parsers } from "./v1/parsers";
-import { binaryParsers as v03Parsers } from "./v03/parsers";
+import { binaryParsers as v1Parsers } from "./v1";
+import { binaryParsers as v03Parsers } from "./v03";
 import { parserByContentType, MappedParser } from "../../parsers";
 import { isString, isBase64, ValidationError, isStringOrObjectOrThrow } from "../../event/validation";
 import CONSTANTS from "../../constants";
