+++ src/transport/http/structured_receiver.ts
@@ -2,8 +2,8 @@ import { CloudEvent, Version } from "../..";
 import { Headers, sanitize } from "./headers";
 import { Parser, JSONParser, MappedParser } from "../../parsers";
 import { parserByContentType } from "../../parsers";
-import { structuredParsers as v1Parsers } from "./v1/parsers";
-import { structuredParsers as v03Parsers } from "./v03/parsers";
+import { structuredParsers as v1Parsers } from "./v1";
+import { structuredParsers as v03Parsers } from "./v03";
 import { isString, isBase64, ValidationError, isStringOrObjectOrThrow } from "../../event/validation";
 import { CloudEventV1, validateV1 } from "../../event/v1";
 import { CloudEventV03, validateV03 } from "../../event/v03";
