--- src/transport/http/v03/index.ts
@@ -1,1 +1,0 @@
-export * from "./headers";
